requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-suit==2.0a1
gunicorn==20.0.4
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
